add: lspのhoverで表示される情報を二重線にする

+++ .bin/.config/nvim/init.lua
@@ -25,7 +25,7 @@ require("lazy").setup("plugins", {
     },
   },
   ui = {
-    border = "single",
+    border = "double",
   },
 })
